use byte if it's a small number

+++ hc-sr04-led/hc-sr04-led.ino
@@ -46,7 +46,7 @@ void setup() {
  * Small blink function that uses the distance for frequency.
  */
 void blink(int distance) {
-  for(int i = 0; i < 5; i++) {
+  for(byte i = 0; i < 5; i++) {
     digitalWrite(LED, HIGH);   // turn the LED on (HIGH is the voltage level)
     delay(min((long)distance, 200));              // wait for a second
     digitalWrite(LED, LOW);    // turn the LED off by making the voltage LOW
